refactor(site/src/pages/AgentsPage): extract AgentChatPage helpers
Pull queue reconciliation, workspace-watch comparisons, and picker option shaping out of AgentChatPage into the modules that already own that state. AgentChatPage drops about 400 lines.

Also write the promoted-queue settle into the messages cache in the same turn as the store apply, so a later queue_update cannot get stomped, and share the right-panel open/width keys from RightPanel instead of duplicating them.
